+++ mdgram.cc
@@ -62,7 +62,7 @@ namespace node_multicast {
 	  mreq.imr_interface.s_addr=htonl(INADDR_ANY);
 
 		if (setsockopt(fd, IPPROTO_IP, IP_ADD_MEMBERSHIP, &mreq, sizeof(mreq)) < 0) {
-			return False();
+			return ThrowException(ErrnoException(errno, "socket"));
 		}
 
 		return True();
@@ -96,7 +96,7 @@ namespace node_multicast {
 	  mreq.imr_interface.s_addr=htonl(INADDR_ANY);
 
 		if (setsockopt(fd, IPPROTO_IP, IP_DROP_MEMBERSHIP, &mreq, sizeof(mreq)) < 0) {
-			return False();
+			return ThrowException(ErrnoException(errno, "socket"));
 		}
 
 		return True();
@@ -113,7 +113,7 @@ namespace node_multicast {
 
 		u_char ttl = args[1]->Int32Value();
 		if (setsockopt(fd, IPPROTO_IP, IP_MULTICAST_TTL, &ttl, sizeof(ttl)) < 0) {
-			return False();
+			return ThrowException(ErrnoException(errno, "socket"));
 		}
 
 		return True();
@@ -138,7 +138,7 @@ namespace node_multicast {
 	  memcpy(ifnamestr, *ifname, ifname.length());
 		ifaddr = inet_addr(ifnamestr);
 		if (setsockopt(fd, IPPROTO_IP, IP_MULTICAST_IF, &ifaddr, sizeof(ifaddr)) < 0) {
-			return False();
+			return ThrowException(ErrnoException(errno, "socket"));
 		}
 
 		return True();