+++ CHANGELOG.markdown
@@ -1,4 +1,5 @@
+NodeJS UDP-Multicast - Change-Log
+=================================
+
 ## Version 0.0.1
-  * Basic functionality: joinGroup / leaveGroup
-
-
+  * Basic functionality: joinGroup / leaveGroup / setMulticastTTL / setMulticastInterface